requirements.txt
faster-whisper>=1.1.0
fastapi>=0.115.0
llama-cpp-python>=0.3.9
numpy>=1.26.0
python-multipart>=0.0.9
PySide6>=6.8.0
sentence-transformers>=3.3.0
sounddevice>=0.5.1
uvicorn>=0.32.0
